DesktopApp/requirements.txt
PyQt5==5.14.2
requests==2.23.0
PyQtWebEngine==5.14.0
pycryptodome==3.9.7
